refactor: :truck: Split header files into single class per header, rename files
