+++ include/system/vfs/file_driver.hpp
@@ -1,42 +1,24 @@
 #pragma once
 
+#include "system/vfs/file_descriptor.hpp"
+
 #include <any>
-#include <compare>
 #include <cstddef>
 #include <cstdint>
 #include <cstdio>
+#include <expected>
 #include <memory>
 #include <span>
 #include <sys/stat.h>
-#include <unistd.h>
-#include <expected>
 #include <system_error>
+#include <unistd.h>
 
 namespace zest::fs {
-
-class FileDescriptor {
-  public:
-    constexpr FileDescriptor(int fd)
-        : m_fd(fd) {}
-
-    explicit operator int() const {
-        return m_fd;
-    }
-
-    std::strong_ordering operator<=>(const FileDescriptor&) const = default;
-
-  private:
-    int m_fd;
-};
-
-inline constexpr FileDescriptor FD_STDIN{0};
-inline constexpr FileDescriptor FD_STDOUT{1};
-inline constexpr FileDescriptor FD_STDERR{2};
-
 class FileDriver : public std::enable_shared_from_this<FileDriver> {
   public:
     virtual void init() = 0;
-    virtual std::expected<FileDescriptor, std::error_condition> open(const char* path, int flags, int mode) = 0;
+    virtual std::expected<FileDescriptor, std::error_condition>
+    open(const char* path, int flags, int mode) = 0;
     virtual ssize_t read(std::any, std::span<std::byte>) = 0;
     virtual int write(std::any, std::span<std::byte>) = 0;
     virtual int close(std::any) = 0;
@@ -48,10 +30,4 @@ class FileDriver : public std::enable_shared_from_this<FileDriver> {
     int32_t update_vfs_entry(FileDescriptor fd, std::any data);
     virtual ~FileDriver() = default;
 };
-
-struct FileEntry {
-  std::shared_ptr<FileDriver> driver;
-  std::any data;
-};
-
 } // namespace zest::fs
