+++ include/system/vfs/usd_driver.hpp
@@ -1,12 +1,15 @@
+#pragma once
+
 #include "pros/rtos.hpp"
-#include "system/vfs/vfs.hpp"
+#include "system/vfs/file_driver.hpp"
 
 namespace zest::fs {
-
 class UsdDriver : public FileDriver {
   public:
     void init() override {}
-    std::expected<FileDescriptor, std::error_condition> open(const char* path, int flags, int mode) override;
+
+    std::expected<FileDescriptor, std::error_condition>
+    open(const char* path, int flags, int mode) override;
     ssize_t read(std::any, std::span<std::byte>) override;
     int write(std::any, std::span<std::byte>) override;
     int close(std::any) override;
@@ -15,7 +18,8 @@ class UsdDriver : public FileDriver {
     off_t lseek(std::any, off_t, int) override;
     int ctl(std::any, const uint32_t, void* const) override;
     ~UsdDriver() override = default;
-private:
+
+  private:
     pros::RecursiveMutex mut{};
 };
